+++ src/parser/read-file.ts
@@ -1,5 +1,4 @@
-import { existsSync, readFileSync } from 'fs';
-import { join as joinPath } from 'path';
+import { existsSync, PathLike, readFileSync } from 'fs';
 
 import { newLineRegex } from '../regex/new-line';
 
@@ -7,13 +6,11 @@ import { newLineRegex } from '../regex/new-line';
  * Read a file into an array of non-empty lines. If no file is found, an empty
  * 	array is returned.
  *
- * @param filename Input filename
+ * @param file Input filename
  * @returns Returns an array of strings, where each non-empty line is an
  * 	element
  */
-export function readFileLines(filename: string): string[] {
-	const file = joinPath(process.cwd(), filename);
-
+export function readFileLines(file: PathLike): string[] {
 	if (existsSync(file)) {
 		return readFileSync(file)
 			.toString()