+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Additions
 - [Issue #12] Add option to prevent reading from process.env
+- [Issue #11] Add support for relative & absolute filepaths
 
 ## [1.1.1]